+++ docs/changelog.rst
@@ -1,6 +1,8 @@
 Changelog
 =========
 
+- :feature:`14` Add :func:`characteristic.immutable` to make certain attributes of classes immutable.
+  Also add ``make_immutable`` argument to :func:`characteristic.attributes`.
 - :feature:`8` Add explicit :class:`characteristic.Attribute` class and use it for default factories.
   Deprecate the ``defaults`` argument of :func:`characteristic.with_init` and :func:`characteristic.attributes`.
 - :release:`0.1.0 <2014-05-11>`